cpu/stm32_common: unified RTC driver
